[BE] Refactor: Clean code
- Clean code
