+++ BE/src/main/java/com/codesquad/baseball06/service/InningService.java
@@ -1,9 +1,9 @@
 package com.codesquad.baseball06.service;
 
 
-import com.codesquad.baseball06.model.Batter;
-import com.codesquad.baseball06.model.Inning;
-import com.codesquad.baseball06.model.Pitcher;
+import com.codesquad.baseball06.model.entity.Batter;
+import com.codesquad.baseball06.model.entity.HalfInning;
+import com.codesquad.baseball06.model.entity.Pitcher;
 import com.codesquad.baseball06.model.type.BattingResult;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -20,10 +20,10 @@ public class InningService {
     this.plateAppearanceService = plateAppearanceService;
   }
 
-  public BattingResult doWork(Inning inning, Pitcher pitcher, Batter batter) {
-    BattingResult plateAppearanceResult = plateAppearanceService.batting(inning, pitcher, batter);
+  public BattingResult doWork(HalfInning halfInning, Pitcher pitcher, Batter batter) {
+    BattingResult plateAppearanceResult = plateAppearanceService.batting(halfInning, pitcher, batter);
 
-    if (inning.isFinished()) {
+    if (halfInning.isFinished()) {
       return BattingResult.END;
     }
